Minor changes & simplifications to Lua interface. 5 new functions.
This changes all typical vector interpolation functions that get
value, time, loopType, pingPong, ease as args from the Lua state;
use unified functions for all value type occurances.
This fixes the slightly annoying problem that it was not possible to
pass a boolean true to pingPong or ease, because lua_tonumber() would
always return 0 in that case.

Also changed:
- obj_getWorldPosition() now takes optional x,y-vector

New:
- obj_getWorldPositionAndRotation() is a more efficient shortcut,
	as often both position and rotation are retrieved together,
	and the underlying matrix chain operation is rather expensive.
- entity_getEmitter()
- entity_getNumEmitters()
- setSceneColor2()
- getSceneColor2()

+++ Aquaria/Entity.h
@@ -356,7 +356,7 @@ public:
 	void revive(float a);
 	void setName(const std::string &name);
 	void doFriction(float dt);
-	void doFriction(float dt, int len);
+	void doFriction(float dt, float len);
 
 	bool isNormalLayer() const
 	{
